+++ src/tools/histogram/vpHistogram.cpp
@@ -1,9 +1,9 @@
 /****************************************************************************
  *
- * $Id: vpHistogram.cpp 3530 2012-01-03 10:52:12Z fspindle $
+ * $Id: vpHistogram.cpp 4303 2013-07-04 14:14:00Z fspindle $
  *
  * This file is part of the ViSP software.
- * Copyright (C) 2005 - 2012 by INRIA. All rights reserved.
+ * Copyright (C) 2005 - 2013 by INRIA. All rights reserved.
  * 
  * This software is free software; you can redistribute it and/or
  * modify it under the terms of the GNU General Public License
@@ -857,7 +857,7 @@ unsigned vpHistogram::sort(std::list<vpHistogramPeak> & peaks)
 
   peaks.sort(compare_vpHistogramPeak);
 
-  return peaks.size();
+  return (unsigned int)peaks.size();
 }
 
 /*!
